ref(streams): Split up consumer implementations
